api/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
langchain==0.1.16
langchain-community==0.0.36
# OpenAI
langchain-openai==0.1.6
openai==1.26
# Cohere
# langchain-cohere==0.1.4
# Math Tool
numexpr==2.8.5
# Weather Tool
openmeteo-requests==1.2.0
requests_cache
retry_requests
numpy
pandas
# pydantic==1.10.8
# Employee handbook tool
pypdf==3.15.1
python-dotenv

